server/requirements.txt
click==8.1.7
Flask==2.3.3
Flask-Cors==4.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pymongo==4.5.0
python-dotenv==1.0.0
six==1.16.0
Werkzeug==2.3.7
requests==2.31.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
